+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Change Log
 [upcoming release] - 2024-..-..
 -------------------------------
 - [FIXED] Increasing geojson precision as the default precision might cause problems with pandahub
+- [ADDED] Add GeographicalRegion and SubGeographicalRegion names and ids to bus df in cim converter
+- [CHANGED] Capitalize first letter of columns busbar_id, busbar_name and substation_id in bus df for cim converter
 
 [2.14.11] - 2024-07-08
 -------------------------------